moved challenge page file
